Implemented a generic liner library.
Needed for I want to use the tests output in the
documentation in the future.
Later it can be a separate project can link content
between your predefined start and end tags, with
lot of freedom.
String file and directory lever linking is implemented.

Even in the same file you have more places to link
the same content it is possible.